+++ testcube_client/settings.py
@@ -111,3 +111,6 @@ class API:
     error = 'errors'
     object_source = 'object_sources'
     result_file = 'result_files'
+    reset_result = 'reset_results'
+    task = 'tasks'
+    run_variable = 'run_variables'